requirements.txt
nicegui==2.22.1
requests==2.32.4
splunk-sdk==2.1.0
python-dotenv==1.1.1